refactor: move `EVMGroupByExec` to its own file
